+++ src/store/booking-options/state.js
@@ -1,6 +1,5 @@
 const state = {
   serviceAvailabilityModel: {},
-  sessions: [],
 }
 
 export default state