+++ src/store/index.js
@@ -2,33 +2,16 @@ import bookings from './bookings'
 import bookingOptions from './booking-options'
 import ui from './ui'
 import settings from './settings'
+import { store as services } from './../modules/services'
 
 const state = {
   app: {
-    bookingsEnabled: false,
     timezone: 'UTC+0',
-    availabilities: {
-      rules: []
-    },
-    sessionLengths: [],
-    displayOptions: {
-      allowCustomerChangeTimezone: false
-    },
-
     bookings: [],
     screenOptionsEndpoint: '',
     statuses: {},
     screenStatuses: []
   },
-}
-
-const getters = {
-  availabilities (state) {
-    return state.app.availabilities.rules
-  },
-  sessionLengths (state) {
-    return state.app.sessionLengths
-  }
 }
 
 const mutations = {
@@ -41,26 +24,8 @@ const mutations = {
     state.bookings.timezone = appState.bookingsTimezone || appState.config.timezone
   },
 
-  setNewAvailabilities (state, rules) {
-    state.app.availabilities.rules = rules
-  },
-
-  setSessionLengths (state, sessionLengths) {
-    state.app.sessionLengths = sessionLengths
-  },
-
   setTimezone (state, value) {
     state.app.timezone = value
-  },
-
-  setDisplayOptions (state, payload) {
-    let updatedOption = {}
-    updatedOption[payload.key] = payload.value
-    state.app.displayOptions = Object.assign({}, state.app.displayOptions, updatedOption)
-  },
-
-  setBookingsEnabled (state, newValue) {
-    state.app.bookingsEnabled = newValue
   },
 
   setNewBookings (state, bookings) {
@@ -78,10 +43,10 @@ export default function ({deepHas, deepSet}) {
       bookings,
       bookingOptions,
       ui,
-      settings
+      settings,
+      services
     },
     state,
-    getters,
     mutations: Object.assign({}, mutations, {
       /**
        * Deep set value of state.