+++ src/transformers/index.js
@@ -1,9 +1,10 @@
 import BookingStoreTransformer from './BookingStoreTransformer'
 import AvailabilityStoreTransformer from './AvilabilityStoreTransformer'
 import AvailabilityReadTransformer from './AvilabilityReadTransformer'
-import StateTransformer from './StateTransformer'
+import ServiceReadTransformer from './ServiceReadTransformer'
 import SessionLengthReadTransformer from './SessionLengthReadTransformer'
 import BookingReadTransformer from './BookingReadTransformer'
+import ServiceStoreTransformer from './ServiceStoreTransformer'
 
 export default function (dependencies) {
   return {
@@ -25,8 +26,11 @@ export default function (dependencies) {
     sessionReadTransformer (container) {
       return new dependencies.bookingWizardComponents.SessionReadTransformer(container.moment)
     },
-    stateTransformer (container) {
-      return new StateTransformer(container.availabilityReadTransformer, container.sessionLengthReadTransformer)
+    serviceReadTransformer (container) {
+      return new ServiceReadTransformer(container.availabilityReadTransformer, container.sessionLengthReadTransformer)
+    },
+    serviceStoreTransformer (container) {
+      return new ServiceStoreTransformer(container.availabilityStoreTransformer)
     },
     transformDatetimeForUi (container) {
       /**